monet: colors: Switch to single-precision floating-point math
The only differences in output that I've noticed with Oklab + gamut
clipping are very small differences in some pure blue (#0000FF) colors,
particularly darker ones. The differences are too small to matter, so
let's switch to 32-bit floats for performance.

+++ app/src/main/java/dev/kdrag0n/android12ext/monet/colors/Cam16UcsLch.kt
@@ -6,9 +6,9 @@ import dev.kdrag0n.android12ext.monet.colors.Lch.Companion.calcLchC
 import dev.kdrag0n.android12ext.monet.colors.Lch.Companion.calcLchH
 
 data class Cam16UcsLch(
-    override val L: Double,
-    override val C: Double,
-    override val h: Double,
+    override val L: Float,
+    override val C: Float,
+    override val h: Float,
 ) : Lch {
     override fun toLinearSrgb() = toCam16Ucs().toLinearSrgb()
